+++ tests/sso_tests/test_signin.py
@@ -12,8 +12,8 @@ class TestSignIn:
         sign_in_payload = SSOMockFactory.make_sign_in_payload(sign_up_payload["phone_number"], sign_up_payload["password"])
         sign_in_response = APIRouter.SSO.sign_in(client, sign_in_payload)
 
-        assert sign_in_response.status_code == 200
-        assert "access_token" in sign_in_response.json()["data"]
+        assert sign_in_response.status_code == 200, f"Ошибка при авторизации пользователя: {sign_in_response.json()}"
+        assert "access_token" in sign_in_response.json()["data"], "Ответ должен содержать access_token"
 
     # Тест авторизации на проверку неверного пароля
     def test_sign_in_wrong_password(self, client: TestClient):
@@ -23,7 +23,7 @@ class TestSignIn:
         wrong_password_payload = SSOMockFactory.make_sign_in_payload(sign_up_payload["phone_number"], "wrong_password")
         sign_in_response = APIRouter.SSO.sign_in(client, wrong_password_payload)
 
-        assert sign_in_response.status_code == 401
+        assert sign_in_response.status_code == 401, f"Пользователь должен получить ошибку, что пароль не верный: {sign_in_response.json()}"
 
     # Тест авторизации на проверку неверного номера телефона
     def test_sign_in_wrong_phone_number(self, client: TestClient):
@@ -33,4 +33,4 @@ class TestSignIn:
         wrong_phone_number_payload = SSOMockFactory.make_sign_in_payload("12312312", sign_up_payload["password"])
         sign_in_response = APIRouter.SSO.sign_in(client, wrong_phone_number_payload)
 
-        assert sign_in_response.status_code == 404
+        assert sign_in_response.status_code == 404, f"Пользователь должен получить ошибку, что номер телефона не найден: {sign_in_response.json()}"
