+++ tests/users/test_user_info.py
@@ -1,12 +1,12 @@
 from fastapi.testclient import TestClient
 
 from tests.APIRouter import APIRouter
-from tests.utils.sso_utils import AuthenticatedUser, sign_up_user
+from tests.utils.flows.SSOFlow import SSOTestFlow, AuthenticatedUser
 
 class TestUserInfo:
     # Тест на получение информации о пользователе
     def test_get_user_info(self, client: TestClient):
-        auth_data = sign_up_user(client)
+        auth_data = SSOTestFlow.sign_up_user(client)
         response = APIRouter.Users.current_user(client, auth_data.headers)
 
         assert response.status_code == 200, f"Ошибка: {response.json()}"
@@ -20,7 +20,7 @@ class TestUserInfo:
 
     def test_public_user_info(self, client: TestClient):
         # Попытка получить публичную информацию о пользователе
-        auth_data: AuthenticatedUser = sign_up_user(client)
+        auth_data: AuthenticatedUser = SSOTestFlow.sign_up_user(client)
 
         user_id = auth_data.user_id
         public_response = APIRouter.Users.public_user_info(client, user_id)
@@ -30,7 +30,7 @@ class TestUserInfo:
 
     def test_private_data_in_public_info(self, client: TestClient):
         # Проверка, что приватные данные не возвращаются в публичной информации
-        auth_data: AuthenticatedUser = sign_up_user(client)
+        auth_data: AuthenticatedUser = SSOTestFlow.sign_up_user(client)
 
         user_id = auth_data.user_id
         public_response = APIRouter.Users.public_user_info(client, user_id)
